+++ src/main/java/http/util/HttpRequestParser.java
@@ -3,13 +3,13 @@ package http.util;
 import http.common.HttpHeaders;
 import http.common.HttpMethod;
 import http.common.URL;
+import http.exception.BadRequestException;
 import http.request.HttpRequest;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
 
-import static http.util.HttpRequestReader.*;
 public class HttpRequestParser {
     private HttpRequestParser() {}
 
@@ -26,5 +26,27 @@ public class HttpRequestParser {
         } catch (IOException e) {
             throw new RuntimeException("잘못된 형식의 요청입니다.");
         }
+    }
+
+    public static String readRequestLine(BufferedReader br) throws IOException {
+        return br.readLine();
+    }
+
+    public static String readHeader(BufferedReader br) throws IOException {
+        String line;
+        StringBuilder strOfHeaders = new StringBuilder();
+        while (!(line = br.readLine()).equals("")) {
+            strOfHeaders.append(line).append("\n");
+        }
+        return strOfHeaders.toString();
+    }
+
+    public static String readBody(BufferedReader br, int contentLength) throws IOException {
+        char[] data = new char[contentLength];
+        int read = br.read(data);
+        if (read == -1) {
+            throw new BadRequestException("요청 형식이 잘못되었습니다.");
+        }
+        return String.valueOf(data);
     }
 }
